fastapi-auth/requirements-minimal.txt
# Minimal requirements for local development
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.0
python-dotenv==1.0.0